Split AppSignal sink into separate modules
As per review feedback: split the new sink style into separate module
files.
